requirements.txt
python-telegram-bot==11.1.0
python-dotenv==0.20.0
redis==4.3.4